Add TrustedSystem.ClearFiles()
The problem is that we previously had a way to add files, but no way to
remove (or even see what files were in the system).  This was preventing
us from changing the system's files on-the-fly.

This solution is not very elegant, but it gets the job done for now.