Replace `assertItemsEqual` with `assertCountEqual`
This, from `six`.  See the root cause that made this necessary:
https://bugs.python.org/issue17866

The one test change that isn't a straight-up "search-replace-reformat"
is because the test _should_ have been `assertItemsEqual`, but wasn't.
It was too strict: it tested the order of items whose order we don't
care about.  And, that test started failing in python 3 for reasons
still obscure to me.  So, I just made it what it always should have
been.

Burndown: from 22 errors down to 2.

Helps #3.