Close file object for `/dev/null`
This is pretty straightforward: I straight-up used a bare `open` and
never closed the file.  Oops!

If this were python3-only, I could use `subprocess.DEVNULL`.  But, that
doesn't work on python2.

Also did some reflow on function calls as needed.

Note that at this point, the output for running all the tests under
python2 or python3 is essentially identical (save only the `b` prefix
for the encoded bytestrings, the name of the tempfile, and the execution
time).

Cleanup for #3.