Stop printing debug messages for `File`
Right now, we have a `try`-`catch` structure, which turns two exceptions
into `print` statements.

First, if we get `IOError`, meaning the file isn't found.  Well, this is
an actually exceptional condition---so, instead of printing a message,
we should let the exception propagate.

Second, if we get `TypeError`, which is a cryptic way to refer to
`file_name` being `None`.  This is not an exceptional condition, and we
shouldn't use exceptions for control flow.  We don't need to print
anything out when we're using a dummy file; this is pretty much only for
tests, and it makes the tests noisy.  Replacing the `try` with an
`if file_name` is just as clear.

Cleanup after #3.