Remove irrelevant assertions
These tests were testing More Than One Thing, by depending on the
details of the recurring task algorithm.  Really, what's relevant for
these tests is that the task is invisible at the moment of checkoff,
invisible at the moment before the new interval, and visible (due) at
the moment the new interval begins.

Cleanup for #7.